backend/requirements.txt
fastapi
uvicorn
opencv-python
pyzbar
google-generativeai
openpyxl
pillow
qrcode
numpy
python-multipart
